+++ ecs/exchange_gen.go
@@ -14,6 +14,7 @@ type Exchange1[A any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange1]. It is safe to call on `nil` instance.
@@ -32,6 +33,7 @@ func NewExchange1[A any](world *World) *Exchange1[A] {
 	return &Exchange1[A]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -62,7 +64,7 @@ func (ex *Exchange1[A]) Add(entity Entity, a *A, rel ...Relation) {
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange1[A]) AddFn(entity Entity, fn func(a *A), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -95,7 +97,7 @@ func (ex *Exchange1[A]) Exchange(entity Entity, a *A, rel ...Relation) {
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange1[A]) ExchangeFn(entity Entity, fn func(a *A), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -152,7 +154,7 @@ func (ex *Exchange1[A]) ExchangeBatchFn(batch *Batch, fn func(entity Entity, a *
 }
 
 func (ex *Exchange1[A]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {
@@ -197,6 +199,7 @@ type Exchange2[A any, B any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange2]. It is safe to call on `nil` instance.
@@ -214,6 +217,7 @@ func NewExchange2[A any, B any](world *World) *Exchange2[A, B] {
 	return &Exchange2[A, B]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -245,7 +249,7 @@ func (ex *Exchange2[A, B]) Add(entity Entity, a *A, b *B, rel ...Relation) {
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange2[A, B]) AddFn(entity Entity, fn func(a *A, b *B), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -279,7 +283,7 @@ func (ex *Exchange2[A, B]) Exchange(entity Entity, a *A, b *B, rel ...Relation) 
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange2[A, B]) ExchangeFn(entity Entity, fn func(a *A, b *B), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -338,7 +342,7 @@ func (ex *Exchange2[A, B]) ExchangeBatchFn(batch *Batch, fn func(entity Entity, 
 }
 
 func (ex *Exchange2[A, B]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A, b *B), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {
@@ -388,6 +392,7 @@ type Exchange3[A any, B any, C any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange3]. It is safe to call on `nil` instance.
@@ -408,6 +413,7 @@ func NewExchange3[A any, B any, C any](world *World) *Exchange3[A, B, C] {
 	return &Exchange3[A, B, C]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -440,7 +446,7 @@ func (ex *Exchange3[A, B, C]) Add(entity Entity, a *A, b *B, c *C, rel ...Relati
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange3[A, B, C]) AddFn(entity Entity, fn func(a *A, b *B, c *C), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -475,7 +481,7 @@ func (ex *Exchange3[A, B, C]) Exchange(entity Entity, a *A, b *B, c *C, rel ...R
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange3[A, B, C]) ExchangeFn(entity Entity, fn func(a *A, b *B, c *C), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -536,7 +542,7 @@ func (ex *Exchange3[A, B, C]) ExchangeBatchFn(batch *Batch, fn func(entity Entit
 }
 
 func (ex *Exchange3[A, B, C]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A, b *B, c *C), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {
@@ -589,6 +595,7 @@ type Exchange4[A any, B any, C any, D any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange4]. It is safe to call on `nil` instance.
@@ -610,6 +617,7 @@ func NewExchange4[A any, B any, C any, D any](world *World) *Exchange4[A, B, C, 
 	return &Exchange4[A, B, C, D]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -643,7 +651,7 @@ func (ex *Exchange4[A, B, C, D]) Add(entity Entity, a *A, b *B, c *C, d *D, rel 
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange4[A, B, C, D]) AddFn(entity Entity, fn func(a *A, b *B, c *C, d *D), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -679,7 +687,7 @@ func (ex *Exchange4[A, B, C, D]) Exchange(entity Entity, a *A, b *B, c *C, d *D,
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange4[A, B, C, D]) ExchangeFn(entity Entity, fn func(a *A, b *B, c *C, d *D), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -742,7 +750,7 @@ func (ex *Exchange4[A, B, C, D]) ExchangeBatchFn(batch *Batch, fn func(entity En
 }
 
 func (ex *Exchange4[A, B, C, D]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A, b *B, c *C, d *D), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {
@@ -798,6 +806,7 @@ type Exchange5[A any, B any, C any, D any, E any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange5]. It is safe to call on `nil` instance.
@@ -820,6 +829,7 @@ func NewExchange5[A any, B any, C any, D any, E any](world *World) *Exchange5[A,
 	return &Exchange5[A, B, C, D, E]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -854,7 +864,7 @@ func (ex *Exchange5[A, B, C, D, E]) Add(entity Entity, a *A, b *B, c *C, d *D, e
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange5[A, B, C, D, E]) AddFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -891,7 +901,7 @@ func (ex *Exchange5[A, B, C, D, E]) Exchange(entity Entity, a *A, b *B, c *C, d 
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange5[A, B, C, D, E]) ExchangeFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -956,7 +966,7 @@ func (ex *Exchange5[A, B, C, D, E]) ExchangeBatchFn(batch *Batch, fn func(entity
 }
 
 func (ex *Exchange5[A, B, C, D, E]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A, b *B, c *C, d *D, e *E), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {
@@ -1015,6 +1025,7 @@ type Exchange6[A any, B any, C any, D any, E any, F any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange6]. It is safe to call on `nil` instance.
@@ -1038,6 +1049,7 @@ func NewExchange6[A any, B any, C any, D any, E any, F any](world *World) *Excha
 	return &Exchange6[A, B, C, D, E, F]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -1073,7 +1085,7 @@ func (ex *Exchange6[A, B, C, D, E, F]) Add(entity Entity, a *A, b *B, c *C, d *D
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange6[A, B, C, D, E, F]) AddFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E, f *F), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -1111,7 +1123,7 @@ func (ex *Exchange6[A, B, C, D, E, F]) Exchange(entity Entity, a *A, b *B, c *C,
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange6[A, B, C, D, E, F]) ExchangeFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E, f *F), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -1178,7 +1190,7 @@ func (ex *Exchange6[A, B, C, D, E, F]) ExchangeBatchFn(batch *Batch, fn func(ent
 }
 
 func (ex *Exchange6[A, B, C, D, E, F]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A, b *B, c *C, d *D, e *E, f *F), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {
@@ -1240,6 +1252,7 @@ type Exchange7[A any, B any, C any, D any, E any, F any, G any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange7]. It is safe to call on `nil` instance.
@@ -1264,6 +1277,7 @@ func NewExchange7[A any, B any, C any, D any, E any, F any, G any](world *World)
 	return &Exchange7[A, B, C, D, E, F, G]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -1300,7 +1314,7 @@ func (ex *Exchange7[A, B, C, D, E, F, G]) Add(entity Entity, a *A, b *B, c *C, d
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange7[A, B, C, D, E, F, G]) AddFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E, f *F, g *G), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -1339,7 +1353,7 @@ func (ex *Exchange7[A, B, C, D, E, F, G]) Exchange(entity Entity, a *A, b *B, c 
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange7[A, B, C, D, E, F, G]) ExchangeFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E, f *F, g *G), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -1408,7 +1422,7 @@ func (ex *Exchange7[A, B, C, D, E, F, G]) ExchangeBatchFn(batch *Batch, fn func(
 }
 
 func (ex *Exchange7[A, B, C, D, E, F, G]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A, b *B, c *C, d *D, e *E, f *F, g *G), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {
@@ -1473,6 +1487,7 @@ type Exchange8[A any, B any, C any, D any, E any, F any, G any, H any] struct {
 	ids       []ID
 	remove    []ID
 	relations []RelationID
+	mask      bitMask
 }
 
 // New creates a new [Exchange8]. It is safe to call on `nil` instance.
@@ -1498,6 +1513,7 @@ func NewExchange8[A any, B any, C any, D any, E any, F any, G any, H any](world 
 	return &Exchange8[A, B, C, D, E, F, G, H]{
 		world: world,
 		ids:   ids,
+		mask:  newMask(ids...),
 	}
 }
 
@@ -1535,7 +1551,7 @@ func (ex *Exchange8[A, B, C, D, E, F, G, H]) Add(entity Entity, a *A, b *B, c *C
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange8[A, B, C, D, E, F, G, H]) AddFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E, f *F, g *G, h *H), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, nil, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -1575,7 +1591,7 @@ func (ex *Exchange8[A, B, C, D, E, F, G, H]) Exchange(entity Entity, a *A, b *B,
 //
 // ⚠️ Do not store the obtained pointers outside of the current context!
 func (ex *Exchange8[A, B, C, D, E, F, G, H]) ExchangeFn(entity Entity, fn func(a *A, b *B, c *C, d *D, e *E, f *F, g *G, h *H), rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 	ex.world.exchange(entity, ex.ids, ex.remove, ex.relations)
 	if fn != nil {
 		ex.runCallback(entity, fn)
@@ -1646,7 +1662,7 @@ func (ex *Exchange8[A, B, C, D, E, F, G, H]) ExchangeBatchFn(batch *Batch, fn fu
 }
 
 func (ex *Exchange8[A, B, C, D, E, F, G, H]) exchangeBatchFn(batch *Batch, fn func(entity Entity, a *A, b *B, c *C, d *D, e *E, f *F, g *G, h *H), remove bool, rel ...Relation) {
-	ex.relations = relations(rel).toRelations(ex.world, ex.ids, ex.relations, 0)
+	ex.relations = relations(rel).toRelations(ex.world, &ex.mask, ex.ids, ex.relations, 0)
 
 	var process func(tableID tableID, start, len int)
 	if fn != nil {